+++ gulpfile.js
@@ -1,5 +1,18 @@
 const { src, dest, parallel } = require( 'gulp' );
+const autoprefixer = require( 'autoprefixer' );
+const cssnano = require( 'cssnano' );
+const postcss = require( 'gulp-postcss' );
 const readme = require( 'gulp-readme-to-markdown' );
+
+function css() {
+	const plugins = [
+		autoprefixer(),
+		cssnano(),
+	];
+	return src( './src/*.css' )
+		.pipe( postcss( plugins ) )
+		.pipe( dest( './build' ) );
+}
 
 function readmeToMarkdown() {
 	return src( [ 'readme.txt' ] )
@@ -14,5 +27,6 @@ function readmeToMarkdown() {
 		.pipe( dest( '.' ) );
 }
 
+exports.css = css;
 exports.readme = readmeToMarkdown;
-exports.default = parallel( readmeToMarkdown );
+exports.default = parallel( css );